+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Disabled by default the storage of celery results - #4169 by @NyanKiyoshi
 - Fix typos and messages in Dashboard 2.0 - #4168 by @benekex2
 - Rename menu items in Dashboard 2.0 - #4172 by @benekex2
+- Adding a non existent variant to a checkout no longer crashes - #4166 by @NyanKiyoshi
 
 ## 2.6.0